authentication/docker/requirements.txt
python-keycloak
pandas
requests